% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{aes_percentage}
\alias{aes_percentage}
\title{Generate mapping for labeling percentages}
\usage{
aes_percentage(relative_to, digits = 0, sep = "")
}
\arguments{
\item{relative_to}{defines proportion that should be calculated, relative to \code{'intersection'}, \code{'group'}, or \code{'all'} observed values}

\item{digits}{number of digits to show (default=0)}

\item{sep}{separator separator between the digit and percent sign (no separator by default)}
}
\description{
Generate mapping for labeling percentages
}
