% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{arrange_venn}
\alias{arrange_venn}
\title{Arrange points for Venn diagram}
\usage{
arrange_venn(
  data,
  sets = NULL,
  radius = 1.5,
  max_iterations = 10,
  verbose = FALSE,
  outwards_adjust = 1.3,
  extract_sets = FALSE,
  extract_regions = FALSE,
  repeat_in_intersections = FALSE,
  starting_grid_size = "auto"
)
}
\arguments{
\item{data}{a dataframe including binary columns representing membership in sets}

\item{sets}{vector with names of columns representing membership in sets}

\item{radius}{the radius of the circle}

\item{max_iterations}{the maximal number of iterations}

\item{verbose}{should debugging notes be printed?}

\item{outwards_adjust}{the multiplier defining the distance from the centre}

\item{extract_sets}{should only sets be extracted?}

\item{extract_regions}{should all unique regions be extracted?}

\item{repeat_in_intersections}{repeat intersection k times where k is the number of sets it belongs to?}

\item{starting_grid_size}{the starting size of the grid for placement of elements}
}
\description{
Arrange points for Venn diagram
}
