% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{compare_between_intersections}
\alias{compare_between_intersections}
\title{Compare covariates between intesections}
\usage{
compare_between_intersections(
  data,
  intersect,
  test = kruskal.test,
  tests = list(),
  ignore = list(),
  ...
)
}
\description{
Compare covariates between intesections
}
