% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_upset_abc_example}
\alias{create_upset_abc_example}
\title{Create an example dataset with three sets: A, B and C}
\usage{
create_upset_abc_example()
}
\description{
Create an example dataset with three sets: A, B and C
}
