% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{geom_venn_region}
\alias{geom_venn_region}
\title{Region of Venn diagram}
\usage{
geom_venn_region(data, mapping = aes_(), sets = NULL, resolution = 250, ...)
}
\arguments{
\item{data}{a dataframe including binary columns representing membership in sets}

\item{mapping}{the aesthetics mapping}

\item{sets}{vector with names of columns representing membership in sets}

\item{resolution}{the resolution of the circle rasterizer}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:geom_polygon]{ggplot2::geom_polygon}}
  \describe{
    \item{\code{stat}}{The statistical transformation to use on the data for this
layer, as a string.}
    \item{\code{position}}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}
    \item{\code{rule}}{Either \code{"evenodd"} or \code{"winding"}. If polygons with holes are
being drawn (using the \code{subgroup} aesthetic) this argument defines how the
hole coordinates are interpreted. See the examples in \code{\link[grid:grid.path]{grid::pathGrob()}} for
an explanation.}
    \item{\code{na.rm}}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
    \item{\code{show.legend}}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}
    \item{\code{inherit.aes}}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
  }}
}
\description{
Region of Venn diagram
}
