% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{get_size_mode}
\alias{get_size_mode}
\title{Retrieve symbol for given mode that can be used in aes mapping with double bang (!!)}
\usage{
get_size_mode(mode)
}
\arguments{
\item{mode}{the mode to use. Accepted values: \code{exclusive_intersection} (alias \code{distinct}), \code{inclusive_intersection} (alias \code{intersect}), \code{inclusive_union} (alias \code{union}), \code{exclusive_union}.}
}
\description{
Retrieve symbol for given mode that can be used in aes mapping with double bang (!!)
}
