% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{intersection_matrix}
\alias{intersection_matrix}
\title{Prepare layers for sets sizes plot}
\usage{
intersection_matrix(
  geom = geom_point(size = 3),
  segment = geom_segment(),
  outline_color = list(active = "black", inactive = "grey70")
)
}
\arguments{
\item{geom}{a geom_point call, allowing to specify parameters (e.g. \code{geom=geom_point(shape='square')})}

\item{segment}{a geom_segment call, allowing to specify parameters (e.g. \code{segment=geom_segment(linetype='dotted')})}

\item{outline_color}{a named list with two colors for outlines of active and inactive dots}
}
\description{
Prepare layers for sets sizes plot
}
