% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{intersection_ratio}
\alias{intersection_ratio}
\title{Barplot annotation of relative intersections sizes}
\usage{
intersection_ratio(
  counts = TRUE,
  bar_number_threshold = 0.75,
  text_colors = c(on_background = "black", on_bar = "white"),
  text = list(),
  text_aes = aes_string(),
  aest = aes_string()
)
}
\description{
A large intersection size can be driven by a large number of members in a group;
to account for that, one can divide the intersection size by the size of a union of the same groups.
This cannot be calculated for the null intersection (observations which do not belong to either of the groups).
}
