% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{intersection_size}
\alias{intersection_size}
\title{Barplot annotation of intersections sizes}
\usage{
intersection_size(
  counts = TRUE,
  bar_number_threshold = 0.85,
  text_colors = c(on_background = "black", on_bar = "white"),
  text = list(),
  text_aes = aes_string(),
  aest = aes_string()
)
}
\description{
Barplot annotation of intersections sizes
}
