% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{reverse_log_trans}
\alias{reverse_log_trans}
\title{Logarithmic scale for use with \code{upset_set_size()}}
\usage{
reverse_log_trans(base = 10)
}
\arguments{
\item{base}{logarithm base (default 10)}
}
\description{
Inspired by \href{https://stackoverflow.com/a/11054781}{Brian Diggs' answer} which is CC-BY-SA 4.0.
}
