% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_default_themes}
\alias{upset_default_themes}
\title{Return the default UpSet themes modified with specified arguments}
\usage{
upset_default_themes(...)
}
\arguments{
\item{...}{arguments passed to `theme()`}
}
\description{
Return the default UpSet themes modified with specified arguments
}
