% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_mode}
\alias{upset_mode}
\title{Layer defining the intersection mode for the data to be displayed}
\usage{
upset_mode(mode)
}
\arguments{
\item{mode}{region selection mode, defines which mode data will be made available for the annotation. See \code{get_size_mode()} for accepted values.}
}
\description{
By default the annotations are given data corresponding to the same mode as the mode of the passed in the \code{upset()} call.
}
