% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_set_size}
\alias{upset_set_size}
\title{Prepare layers for sets sizes plot}
\usage{
upset_set_size(
  geom = geom_bar(width = 0.6),
  position = "left",
  mapping = aes()
)
}
\arguments{
\item{geom}{a geom to use}

\item{position}{on which side of the plot should the set sizes be displayed ('left' or 'right')}

\item{mapping}{additional aesthetics}
}
\description{
Prepare layers for sets sizes plot
}
