% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{upset_test}
\alias{upset_test}
\title{Test for differences between intersections}
\usage{
upset_test(data, intersect, comparisons = list())
}
\description{
This is a wrapper around \code{compare_between_intersections()}, adding sorting by fdr, warnings, etc.
}
