% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\docType{data}
\name{upset_themes}
\alias{upset_themes}
\title{List of default themes for upset components}
\format{
An object of class \code{list} of length 4.
}
\usage{
upset_themes
}
\description{
List of default themes for upset components
}
\keyword{datasets}
