% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{intersection_ratio}
\alias{intersection_ratio}
\title{Barplot annotation of relative intersections sizes}
\usage{
intersection_ratio(
  counts = TRUE,
  bar_number_threshold = 0.75,
  text_colors = c(on_background = "black", on_bar = "white"),
  text = list(),
  text_aes = aes_string(),
  aest = aes_string()
)
}
\description{
sometimes the large intersection size is driven by the large number of members in a group
to account for that, one can divide the intersection size by the union size of the same groups
obviosuly, this canot be calculated for the null intersection (observations which do not belong to either of the groups)
}
