% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset}
\alias{upset}
\title{Compose an UpSet plot}
\usage{
upset(
  data,
  intersect,
  base_annotations = list(`Intersection size` = intersection_size(counts = TRUE)),
  name = "group",
  annotations = list(),
  themes = upset_themes,
  stripes = upset_stripes,
  labeller = identity,
  height_ratio = 0.5,
  width_ratio = 0.3,
  wrap = FALSE,
  overall_sizes = TRUE,
  overall_sizes_bar_width = 0.6,
  queries = list(),
  dot_size = 3,
  ...
)
}
\arguments{
\item{data}{a dataframe including binary columns representing membership in classes}

\item{intersect}{which columns should be used to compose the intersection}

\item{stripes}{a two-element characters vector, specifying the background colors for odd and even rows}

\item{labeller}{function modifying the names of the sets (rows in the matrix)}

\item{height_ratio}{ratio of the intersection matrix to intersection size height}

\item{width_ratio}{ratio of the overall set size width to intersection matrix width}

\item{overall_sizes}{whether to show the overall set sizes (barplot to the left), default TRUE}

\item{overall_sizes_bar_width}{the thickness of the bars in the overal set sizes barplot}

\item{queries}{a list of queries generated with `upset_query()`}

\item{dot_size}{size of the points on the intersection matrix}

\item{...}{passed to upset_data() which accepts: `(min_size=0, keep_empty_groups=FALSE, warn_when_dropping_groups=TRUE, sort_sets='descending', sort_intersections='descending')`}

\item{sort_sets}{whether to sort the rows in the intersection matrix (descending sort by default); one of: `'ascending'`, `'descending'`, `FALSE`}

\item{sort_intersections}{whether to sort the columns in the intersection matrix (descending sort by default); one of: `'ascending'`, `'descending'`, `FALSE`}
}
\description{
Compose an UpSet plot
}
