% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_annotate}
\alias{upset_annotate}
\title{Shorthand for annotations creation, using prespecified aes(x=intersection)}
\usage{
upset_annotate(y, geom)
}
\arguments{
\item{y}{A geom to be used as an annotation}
}
\description{
Shorthand for annotations creation, using prespecified aes(x=intersection)
}
