% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_annotate}
\alias{upset_annotate}
\title{Annotation panel shorthand}
\usage{
upset_annotate(y, geom)
}
\arguments{
\item{y}{A string with the name of the y aesthetic}

\item{geom}{A geom to be used as an annotation}
}
\description{
Simplifies creation of annotation panels, automatically building aesthetics mappings,
at a cost of lower flexibility than when providing a custom mapping; \code{aes(x=intersection)} is prespecified.
}
