% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_default_themes}
\alias{upset_default_themes}
\title{Default themes modified by specified arguments}
\usage{
upset_default_themes(...)
}
\arguments{
\item{...}{arguments passed to \code{theme()}}
}
\description{
Return the default UpSet themes with all themes modified with provided arguments
}
