% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_set_size}
\alias{upset_set_size}
\title{Prepare layers for sets sizes plot}
\usage{
upset_set_size(geom = geom_bar, layers = list(), mapping = aes(), ...)
}
\arguments{
\item{geom}{the geom to use}

\item{layers}{a list of additional layers (scales, geoms) to be included on the plot}

\item{mapping}{additional aesthetics}

\item{...}{passed to the geom}
}
\description{
Prepare layers for sets sizes plot
}
