% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_text_percentage}
\alias{upset_text_percentage}
\title{Generate percentage label of the intersection/union sizes ratio}
\usage{
upset_text_percentage(digits = 0, sep = "")
}
\arguments{
\item{digits}{How many digits to show when rounding the percentage?}

\item{sep}{set to space (\code{' '}) if you prefer a whitespace between the number and the \verb{\\\%} sign.}
}
\description{
For use together with \code{intersection_size} or \code{intersection_ratio}
}
\examples{
ggplot2::aes_(label=upset_text_percentage())
}
